api/requirements.txt
flake8==6.1.0
Django==4.2.6
djangorestframework==3.14.0
psycopg2-binary>=2.9.9
Pillow==10.0.0
pytest-django==4.5.2
pytest-mock==3.12.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.0
django-filter==23.3
drf-standardized-errors[openapi]==0.12.5
pytest-cov==4.1.0
django-cors-headers==4.3.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cleanup==8.0.0